% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r
\name{decisions}
\alias{decisions}
\title{Show decision log for match decisions}
\usage{
decisions(x)
}
\arguments{
\item{x}{a swiss object, as returned by \code{generate_matches}.}
}
\value{

}
\description{
Show decision log for match decisions
}
