% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_pairings.r
\name{generate_matches}
\alias{generate_matches}
\title{Generate new round of matches}
\usage{
generate_matches(stats, players, seed = NULL)
}
\arguments{
\item{stats}{player statistics, as generated by \code{\link{player_stats}}.}

\item{players}{a character vector, giving players to consider in generating
new rounds. Players that are present in \code{stats}, but absent in
\code{players}, are considered to have dropped out of the tournament early.
Any players in \code{players} that are eliminated are automatically removed
from consideration.}

\item{seed}{a random seed, if desired.}
}
\value{

}
\description{
Generate new round of matches
}
