% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_pairings.r
\name{player_stats}
\alias{player_stats}
\title{Collect player statistics}
\usage{
player_stats(matches, players, elimination)
}
\arguments{
\item{matches}{a data.table, containing match results from previous rounds.
Present columns are round (integer), P1 (character), P2 (character), victor
(integer, 1 or 2).}

\item{players}{a character vector, containing names of players to collect
statistics for.}

\item{elimination}{an integer, indicating the number of losses after which a
player is eliminated. Eliminated players are not considered when deciding
on byes or pairings.}
}
\value{
a list of two elements: \code{player_stats}, giving win/loss/bye
  records for each player; \code{pairing_counts}, giving counts of each
  player matchup, with player one on the row and player two on the column.
}
\description{
Collect player statistics
}
