% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_package.r
\docType{package}
\name{swiss-package}
\alias{swiss}
\alias{swiss-package}
\title{swiss: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
A quick package for handling a Survival Swiss tournament.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
